docker/networksimulator/requirements.txt
jinja2
web3
python-dotenv
